project-01 uva-10062 Accepted

+++ project-01/uva-10062/main.cpp
@@ -26,10 +26,10 @@ int main()
             cout << "\n";
         }
 
-        for (int i = 0; i < s.length(); i++)
+        for (int i = 0; i < (int)s.length(); i++)
         {
             int c = 0;
-            for (int j = 0; j < s.length(); j++)
+            for (int j = 0; j < (int)s.length(); j++)
             {
                 if (s[i] == s[j])
                 {
@@ -40,10 +40,10 @@ int main()
             pair.push_back(p);
         }
 
-        for (int i = 0; i < s.length(); i++)
+        for (int i = 0; i < (int)s.length(); i++)
         {
             int c = 0;
-            for (int j = 0; j < s.length(); j++)
+            for (int j = 0; j < (int)s.length(); j++)
             {
                 if (s[i] == s[j])
                 {
@@ -67,7 +67,7 @@ int main()
                    }),
                    end(pair));
 
-        for (int i = 0; i < pair.size(); i++)
+        for (int i = 0; i < (int)pair.size(); i++)
         {
             cout << pair[i].ascii << " " << pair[i].value << "\n";
         }
